caching: Use separate caches for config variants and other contexts.
* The `_cache` directory is now organized in multiple "sub-caches" for
  different contexts.
* A new context is created when config variants or overrides are applied.
* `serve` context uses a different context that the other commends, to prevent
  the `bake` command's output from messing up the preview server (e.g. with
  how asset URLs are generated differently between the two).
* Fix a few places where the cache directory was referenced directly.

+++ piecrust/commands/builtin/util.py
@@ -59,8 +59,8 @@ class PurgeCommand(ChefCommand):
         pass
 
     def run(self, ctx):
-        cache_dir = ctx.app.cache_dir
-        if os.path.isdir(cache_dir):
+        cache_dir = ctx.app.sub_cache_dir
+        if cache_dir and os.path.isdir(cache_dir):
             logger.info("Purging cache: %s" % cache_dir)
             shutil.rmtree(cache_dir)
